cpu6: Various fixes and revised call logic
The fixes deal with various wrong way round registers and the like plus
some other small decode side fun.

The big one is the call/ret handling. From trying to run the code it seems
that the actual behaviour is a hybrid bra/link and stack.

CALL seems to do
	st rt,(-sp)
	mov rt, pc

RET
	mov pc, rt
	ld rt,(sp+)

so it's a stack of call/returns but the top fo the stack is RT