cpu6: flags
This gets us a lot closer. We can't yet pass the CPU self test but sanity is
starting to occur.

The logic appears to be that 16-19 are signed branches. However because they
are based around a subtraction model in order to make the same branch
mnemonics and behaviour consistent with the case of a move or load all the
load operations set N according to the loaded to bit and set V regardless.

This means you can sensibly write both

                     ld b,#n
                     sub a
                     blt

and

                     ld b,(foo)
                     blt