cpu6: correct load and a maybe sll16
Loads do not in fact set N appropriately and set V, they clear N and set
V based on the inverse of the top bit.

It's an odd approach but would make logical sense if the older CPU didn't
set N or have the signed maths and they wanted to make the signed branches
work without breaking existing code.

Anyway it fits the CPU test pattern so we are now up to the shifts.