centurion: big updates
- Processor model based on the EE200 documentation someone discovered, and
  which is the basis of the CPU6
- Fill out a lot of the I/O card interfaces

With this done we fail one of the I/O tests as the memory emulated on the
card is too small, and the MMU test as we simply don't know enough about the
MMU at this point.

Alas until we recover more information there's not enough data to fill out
the disk interfaces other than the Hawk as we don't really understand the
0x41/0x43 interactions and some of the other DMA command patterns or have
any samples of a disk write.