cpu6: further work on flags
We seem to have inc and add sane now. The sub side passes but is suspect.

More of a problem is the pattern used to shuffle and sub using C. This seems
to imply that C is the live ALU flags low and pri high but the pri cannot
be written perhaps ?