mux: Improve interrupt emulation
RX interrupt will be re-triggered if the IRQ handler didn't service input. This
is done for preventing IRQ losses because in @LOAD the same IPL 6 is used for
both rx and tx, but only one of these events is serviced per interrupt.

CTS is hardcoded to 1 in order to allow the output routine to run