mux: Fix a small address decoding bug
UARTs occupy offsets 0 - 7, common control block starts at 8, and in fact the
CRT driver in @LOAD also pokes offset 8 for yet unknown reason. Previously this
offset was aliased to 0, which is a configuration register, responsible for baud
rate. It only worked because we currently don't handle the baud rate in any way